scripts/requirements.txt
# Dependencies for scripts (ingest_*.py, fuzzy_match.py, test_ingestion.py)
python-dotenv>=1.0.0
requests>=2.28.0
supabase>=2.0.0
rapidfuzz>=3.0.0
pytest>=7.0.0
